backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai>=1.40.0
pydantic
httpx
